+++ src/ExceptionHandlerHelper.php
@@ -14,9 +14,8 @@ namespace MarcinOrlowski\ResponseBuilder;
  */
 
 use Exception;
-use Illuminate\Foundation\Exceptions\Handler as ExceptionHandler;
 use Symfony\Component\HttpFoundation\Response;
-use Config;
+use Illuminate\Support\Facades\Config;
 
 /**
  * Class ExceptionHandlerHelper
@@ -26,39 +25,37 @@ class ExceptionHandlerHelper
 	/**
 	 * Render an exception into an HTTP response.
 	 *
-	 * @param  \Illuminate\Http\Request $request   Request object
-	 * @param  \Exception               $exception Exception
+	 * @param  \Illuminate\Http\Request $request Request object
+	 * @param  \Exception               $ex      Exception
 	 *
 	 * @return \Illuminate\Http\Response
 	 */
-	public static function render($request, Exception $exception)
+	public static function render($request, Exception $ex)
 	{
-		if ($exception instanceof \Symfony\Component\HttpKernel\Exception\HttpException) {
-			switch ($exception->getStatusCode()) {
+		if ($ex instanceof \Symfony\Component\HttpKernel\Exception\HttpException) {
+			switch ($ex->getStatusCode()) {
 				case Response::HTTP_NOT_FOUND:
-					$result = ResponseBuilder::errorWithHttpCode(Config::get('response_builder.exception_handler.exception.http_not_found'),
-						$exception->getStatusCode());
+					$result = static::error($ex, Config::get('response_builder.exception_handler.exception.http_not_found'));
 					break;
 
 				case Response::HTTP_SERVICE_UNAVAILABLE:
-					$result = ResponseBuilder::errorWithHttpCode(Config::get('response_builder.exception_handler.exception.http_service_unavailable'),
-						$exception->getStatusCode());
+					$result = static::error($ex, Config::get('response_builder.exception_handler.exception.http_service_unavailable'));
 					break;
 
 				default:
-					$msg = trim($exception->getMessage());
+					$msg = trim($ex->getMessage());
 					if ($msg == '') {
-						$msg = 'Exception code #' . $exception->getStatusCode();
+						$msg = 'Exception code #' . $ex->getStatusCode();
 					}
 
-					$result = ResponseBuilder::error(Config::get('response_builder.exception_handler.exception.http_exception'),
-						['message' => $msg], null, $exception->getStatusCode());
+					$result = static::error($ex, Config::get('response_builder.exception_handler.exception.http_exception'),
+						['message' => $msg]);
 					break;
 			}
 		} else {
-			$msg = trim($exception->getMessage());
+			$msg = trim($ex->getMessage());
 			if (Config::get('response_builder.exception_handler.include_class_name', false)) {
-				$class_name = get_class($exception);
+				$class_name = get_class($ex);
 				if ($msg != '') {
 					$msg = $class_name . ': ' . $msg;
 				} else {
@@ -66,11 +63,35 @@ class ExceptionHandlerHelper
 				}
 			}
 
-			$result = ResponseBuilder::error(Config::get('response_builder.exception_handler.exception.uncaught_exception'),
-				['message' => $msg], null, 500);
+			$result = static::error($ex, Config::get('response_builder.exception_handler.exception.uncaught_exception'),
+				['message' => $msg], Response::HTTP_INTERNAL_SERVER_ERROR);
 		}
 
 		return $result;
 	}
 
+	/**
+	 * @param Exception $ex
+	 * @param integer   $error_code
+	 * @param array     $lang_args
+	 * @param integer   $http_code
+	 *
+	 * @return Response
+	 */
+	protected static function error(Exception $ex, $error_code, array $lang_args = [], $http_code = 0)
+	{
+		if ($http_code == 0) {
+			$http_code = $ex->getStatusCode();
+		}
+
+		$data = [];
+		if (Config::get('app.debug')) {
+			$data = [
+				'file' => $ex->getFile(),
+				'line' => $ex->getLine(),
+			];
+		}
+
+		return ResponseBuilder::error($error_code, $lang_args, $data, $http_code);
+	}
 }