+++ CHANGES.md
@@ -5,6 +5,7 @@
 * v1.5.1 (2016-04-21)
    * Added errorWithMessageAndData()
    * Added env('EX_INCLUDE_CLASS_NAME') to control include_class_name exception handler feature
+   * If app runs in DEBUG mode, ExceptionHandlerHelper now adds 'file' and 'line' to returned JSON
 
 * v1.5.0 (2016-04-18)
    * ExHandler: ExceptionHandler is now replaced by ExceptionHandlerHelper